fix test & use [ubuntu-24.04-arm] in Actions

+++ go/src/sign/dilithium/internal/reduction/reduction_test.go
@@ -41,8 +41,8 @@ func TestModRed_MontgomeryMulWithDilithium(t *testing.T) {
 		bm := uint32((uint64(b) * (1 << 32)) % common.Q)
 
 		// Montgomery mul should be equal to Encoding.
-		prod := MontgomeryMul(uint32(am), uint32(bm))
-		expected := uint32((uint64((uint64(a)*uint64(b))%common.Q) * (1 << 32)) % common.Q)
+		prod := MontgomeryMul(am, bm)
+		expected := uint32(((uint64(a) * uint64(b)) % common.Q) * (1 << 32) % common.Q)
 		assert.Equal(t, expected, prod)
 
 		decoded := MontgomeryMul(prod, 1)
